LPD-45310 Create the NumericLocalizedObjectField component, separate the numeric component into NumericBase and NumericLocalizedObjectField
